Arch: Further cleaning of IFC importer - moved aux functions to importIFCHelper
